Fixes FAB being hidden under screen
