+++ vector/src/main/res/layout/fragment_new_home_detail.xml
@@ -78,24 +78,52 @@
 
         </com.google.android.material.appbar.AppBarLayout>
 
+        <androidx.fragment.app.FragmentContainerView
+            android:id="@+id/roomListContainer"
+            android:layout_width="match_parent"
+            android:layout_height="match_parent"
+            app:layout_behavior="@string/appbar_scrolling_view_behavior" />
+
         <androidx.constraintlayout.widget.ConstraintLayout
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            app:layout_behavior="com.google.android.material.appbar.AppBarLayout$ScrollingViewBehavior">
+            android:layout_gravity="bottom|end">
 
-            <androidx.fragment.app.FragmentContainerView
-                android:id="@+id/roomListContainer"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                app:layout_constraintBottom_toTopOf="@id/bottomNavigationView"
-                app:layout_constraintTop_toTopOf="parent" />
-
-            <com.google.android.material.bottomnavigation.BottomNavigationView
-                android:id="@+id/bottomNavigationView"
-                android:layout_width="match_parent"
+            <com.google.android.material.floatingactionbutton.FloatingActionButton
+                android:id="@+id/newLayoutOpenSpacesButton"
+                style="@style/Widget.Vector.FloatingActionButton"
+                android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
+                android:layout_marginBottom="20dp"
+                android:accessibilityTraversalAfter="@id/newLayoutCreateChatButton"
+                android:contentDescription="@string/a11y_open_spaces"
+                android:src="@drawable/ic_open_spaces"
+                android:visibility="visible"
+                app:backgroundTint="?attr/vctr_toolbar_background"
+                app:fabSize="mini"
+                app:layout_constraintBottom_toTopOf="@id/newLayoutCreateChatButton"
+                app:layout_constraintEnd_toEndOf="@id/newLayoutCreateChatButton"
+                app:layout_constraintStart_toStartOf="@id/newLayoutCreateChatButton"
+                app:tint="?attr/colorPrimary"
+                tools:targetApi="lollipop_mr1"
+                tools:visibility="visible" />
+
+
+            <com.google.android.material.floatingactionbutton.FloatingActionButton
+                android:id="@+id/newLayoutCreateChatButton"
+                style="@style/Widget.Vector.FloatingActionButton"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_marginEnd="16dp"
+                android:layout_marginBottom="16dp"
+                android:accessibilityTraversalBefore="@id/roomListView"
+                android:contentDescription="@string/a11y_create_message"
+                android:src="@drawable/ic_new_chat"
+                android:visibility="visible"
                 app:layout_constraintBottom_toBottomOf="parent"
-                app:menu="@menu/home_bottom_navigation" />
+                app:layout_constraintEnd_toEndOf="parent"
+                tools:targetApi="lollipop_mr1"
+                tools:visibility="visible" />
 
         </androidx.constraintlayout.widget.ConstraintLayout>
